+++ BALSAMIC/commands/report/base.py
@@ -2,7 +2,6 @@
 import click
 
 from BALSAMIC.commands.report.deliver import deliver as deliver_command
-from BALSAMIC.commands.report.status import status as status_command
 
 
 @click.group()
@@ -13,4 +12,3 @@ def report(context: click.Context):
 
 
 report.add_command(deliver_command)
-report.add_command(status_command)