+++ BALSAMIC/constants/analysis.py
@@ -8,7 +8,8 @@ from BALSAMIC.constants.cache import DockerContainers
 class LogFile:
     """Logfile constants"""
 
-    LOGNAME: str = "balsamic.log"
+    RUN_LOGNAME: str = "balsamic_run.log"
+    CONFIG_LOGNAME: str = "balsamic_config.log"
 
 
 class SubmitSnakemake:
@@ -32,6 +33,7 @@ class Gender(StrEnum):
 
     FEMALE: str = "female"
     MALE: str = "male"
+    UNKNOWN: str = "unknown"
 
 
 class AnalysisType(StrEnum):