+++ CHANGELOG.rst
@@ -14,13 +14,15 @@ Changed:
 * changed snakemake from running immediate submit to submitting snakemake job to cluster https://github.com/Clinical-Genomics/BALSAMIC/pull/1558
 * Updated ensembl vep to version 113.4 https://github.com/Clinical-Genomics/BALSAMIC/pull/1567
 * Updated rank model to use GNOMADAF_popmax instead of deprecated gnomAD_AF https://github.com/Clinical-Genomics/BALSAMIC/pull/1567
+* given vs predicted sex check changed to warning https://github.com/Clinical-Genomics/BALSAMIC/pull/1581
+* using predicted sex in CNV analysis if assigned sex is unknown https://github.com/Clinical-Genomics/BALSAMIC/pull/1581
 
 Removed:
 ^^^^^^^^
 * removed immediate submit functionality https://github.com/Clinical-Genomics/BALSAMIC/pull/1558
 * removed unused code for benchmark plotting https://github.com/Clinical-Genomics/BALSAMIC/pull/1558
 * removed functionality to disable variant callers https://github.com/Clinical-Genomics/BALSAMIC/pull/1558
-
+* removed balsamic status check functionality https://github.com/Clinical-Genomics/BALSAMIC/pull/1581
 
 Fixed:
 ^^^^^^